+++ programs/standalone/examples/TestRunner/classic-firebug/BasicLoggingObjectOriented.php
@@ -38,6 +38,4 @@ $firephp->fb(array('2 SQL queries took 0.06 seconds',array(
    array('SELECT * FROM Bar','0.04',array('row1','row2'))
   )),FirePHP::TABLE);
 
-/* Will show only in "Server" tab for the request */
-$firephp->fb(apache_request_headers(),'RequestHeaders',FirePHP::DUMP);
-
+$firephp->fb(phpversion(), 'PHP Version', FirePHP::DUMP);