+++ programs/standalone/examples/TestRunner/classic-firebug/BasicLoggingProcedural.php
@@ -31,6 +31,4 @@ fb(array('2 SQL queries took 0.06 seconds',array(
    array('SELECT * FROM Bar','0.04',array('row1','row2'))
   )),FirePHP::TABLE);
 
-/* Will show only in "Server" tab for the request */
-fb(apache_request_headers(),'RequestHeaders',FirePHP::DUMP);
-
+fb(phpversion(), 'PHP Version', FirePHP::DUMP);