+++ programs/standalone/examples/TestRunner/insight-devcomp/snippets/Introduction-Example.php
@@ -6,7 +6,7 @@
 
 define('INSIGHT_IPS', '*');
 define('INSIGHT_AUTHKEYS', '*');
-define('INSIGHT_PATHS', __DIR__);
+define('INSIGHT_PATHS', dirname(__FILE__));
 define('INSIGHT_SERVER_PATH', './_insight_.php');
 require_once('FirePHP/Init.php');